@@ -1,50 +1,47 @@
 name: HackRU
 description: The Official HackRU Flutter App
 version: 1.2.0
-author: HackRU RnD Flutter Team [Dhruvil Patel, Paul John, Mickey J. Winters]
-homepage: https://github.com/HackRU/OneAppFlutter
+homepage: http://hackru.org/
+repository: https://github.com/HackRU/OneAppFlutter
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.13.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
   # App Dependencies from (https://pub.dev/)
-  cupertino_icons: ^0.1.3
-  font_awesome_flutter: ^8.8.1
-  groovin_material_icons: ^1.1.6
-  hidden_drawer_menu: ^2.0.1
+  cupertino_icons: ^1.0.4
+  equatable: ^2.0.3
+  firebase_messaging: ^11.2.11
+  flare_flutter: ^3.0.2
+  flutter_launcher_icons: ^0.9.2
+  flutter_linkify: ^5.0.2
+  flutter_local_notifications: ^9.4.0
+  font_awesome_flutter: ^9.2.0
+  hidden_drawer_menu: ^3.0.1
   #qr_flutter: ^3.2.0
   #qr_code_scanner: ^0.0.12
-  http: ^0.12.2
-  url_launcher: ^5.4.2
-  rubber: ^0.4.0
-  webview_flutter: ^0.3.22+1
-  # pinch_zoom_image_updated: ^0.2.6
-  flare_flutter: ^2.0.5
-  flutter_linkify: ^3.1.3
-  equatable: ^1.2.3
-  meta: ^1.1.8
-  intl: ^0.16.1
-  path_provider: ^1.6.11
-  flutter_launcher_icons: ^0.7.5
-  json_annotation: ^3.0.1
-  provider: ^4.3.1
-  shared_preferences: ^0.5.8
-  flutter_local_notifications: ^1.4.4+4
-  rxdart: ^0.24.1
-  firebase_messaging: ^7.0.0
-#  dart_lcs_client: ^0.0.3
+  http: ^0.13.4
+  intl: ^0.17.0
+  json_annotation: ^4.4.0
+  meta: ^1.7.0
+  path_provider: ^2.0.9
+  provider: ^6.0.2
+  rubber: ^1.0.1
+  rxdart: ^0.27.3
+  shared_preferences: ^2.0.13
+  url_launcher: ^6.0.20
+  webview_flutter: ^3.0.1
 
 dev_dependencies:
-  test: ^1.14.0
+  test: ^1.20.2
   flutter_test:
     sdk: flutter
-  build_runner: ^1.10.0
-  json_serializable: ^3.3.0
-  pedantic: ^1.9.0
+  build_runner: ^2.1.8
+  json_serializable: ^6.1.5
+  pedantic: ^1.11.1
 
 flutter_icons:
   android: true
@@ -52,9 +49,7 @@ flutter_icons:
   image_path: "assets/hackru-logos/hackru_red.png"
 
 flutter:
-  # The following line ensures that the Material Icons font is
-  # included with your application, so that you can use the icons in
-  # the material Icons class.
+  generate: true
   uses-material-design: true
   assets:
   - assets/
