@@ -1,21 +1,20 @@
 import 'package:HackRU/styles.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:url_launcher/url_launcher.dart' as launcher;
 
 class LinkTextSpan extends TextSpan {
-  LinkTextSpan({TextStyle style, String url, String text})
+  LinkTextSpan({TextStyle? style, String? url, String? text})
       : super(
           style: style,
           text: text ?? url,
           recognizer: TapGestureRecognizer()
-            ..onTap = () => launcher.launch(url),
+            ..onTap = () => launcher.launch(url!),
         );
 }
 
 class StringParser extends StatelessWidget {
-  final String text;
+  final String? text;
 
   StringParser({@required this.text});
 
@@ -93,10 +92,10 @@ class StringParser extends StatelessWidget {
       decoration: TextDecoration.lineThrough,
     );
 
-    final words = text.split(' ');
+    final words = text?.split(' ');
     var span = <TextSpan>[];
 
-    words.forEach((word) {
+    words?.forEach((word) {
       if (_isLink(word)) {
         var eWord = word.replaceAll(RegExp(r'[<>]'), '');
         span.add(LinkTextSpan(
@@ -140,7 +139,7 @@ class StringParser extends StatelessWidget {
       );
     } else {
       return Text(
-        text,
+        text!,
         style: TextStyle(
           fontSize: 15.0,
           color: _brightnessValue == Brightness.light ? pink_dark : yellow,
