@@ -1,21 +1,20 @@
 import 'dart:async';
-import 'dart:convert';
-import 'dart:io';
+// import 'dart:convert';
+// import 'dart:io';
 
 import 'package:HackRU/models/cred_manager.dart';
 import 'package:HackRU/styles.dart';
 import 'package:HackRU/services/hackru_service.dart';
 import 'package:HackRU/models/models.dart';
 import 'package:HackRU/ui/pages/dashboard/announcement_card.dart';
-import 'package:HackRU/ui/widgets/dialog/notification_onclick.dart';
+// import 'package:HackRU/ui/widgets/dialog/notification_onclick.dart';
 import 'package:flare_flutter/flare_actor.dart';
-import 'package:flutter/cupertino.dart';
-import 'package:flutter/foundation.dart';
+// import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
-import 'package:HackRU/main.dart';
-import 'package:firebase_messaging/firebase_messaging.dart';
-import 'package:flutter_local_notifications/flutter_local_notifications.dart';
+// import 'package:HackRU/main.dart';
+// import 'package:firebase_messaging/firebase_messaging.dart';
+// import 'package:flutter_local_notifications/flutter_local_notifications.dart';
 
 class Dashboard extends StatefulWidget {
   @override
@@ -26,123 +25,123 @@ class DashboardState extends State<Dashboard> {
   var _displayTimerBanner = true;
   static DateTime cacheTTL = DateTime.now();
 
-  final FirebaseMessaging _firebaseMessaging = FirebaseMessaging();
+  // late final FirebaseMessaging _firebaseMessaging;
 
   @override
   void initState() {
-    setupPushNotifications();
-    _requestIOSPermissions();
-    _configureDidReceiveLocalNotificationSubject();
-    _configureSelectNotificationSubject();
+    // setupPushNotifications();
+    // _requestIOSPermissions();
+    // _configureDidReceiveLocalNotificationSubject();
+    // _configureSelectNotificationSubject();
   }
 
   /// ===========================================================
   ///                     PUSH NOTIFICATIONS
   /// ===========================================================
 
-  void _requestIOSPermissions() {
-    flutterLocalNotificationsPlugin
-        .resolvePlatformSpecificImplementation<
-            IOSFlutterLocalNotificationsPlugin>()
-        ?.requestPermissions(
-          alert: true,
-          badge: true,
-          sound: true,
-        );
-  }
+  // void _requestIOSPermissions() {
+  //   flutterLocalNotificationsPlugin
+  //       .resolvePlatformSpecificImplementation<
+  //           IOSFlutterLocalNotificationsPlugin>()
+  //       ?.requestPermissions(
+  //         alert: true,
+  //         badge: true,
+  //         sound: true,
+  //       );
+  // }
 
-  void _configureDidReceiveLocalNotificationSubject() {
-    didReceiveLocalNotificationSubject.stream
-        .listen((ReceivedNotification receivedNotification) async {
-      await showDialog(
-        context: context,
-        builder: (BuildContext context) => CupertinoAlertDialog(
-          title: receivedNotification.title != null
-              ? Text(receivedNotification.title)
-              : null,
-          content: receivedNotification.body != null
-              ? Text(receivedNotification.body)
-              : null,
-          actions: [
-            CupertinoDialogAction(
-              isDefaultAction: true,
-              child: Text('Ok'),
-              onPressed: () async {
-                print('Not implemented Yet');
-              },
-            )
-          ],
-        ),
-      );
-    });
-  }
+  // void _configureDidReceiveLocalNotificationSubject() {
+  //   didReceiveLocalNotificationSubject.stream
+  //       .listen((ReceivedNotification receivedNotification) async {
+  //     await showDialog(
+  //       context: context,
+  //       builder: (BuildContext context) => CupertinoAlertDialog(
+  //         title: receivedNotification.title != null
+  //             ? Text(receivedNotification.title)
+  //             : null,
+  //         content: receivedNotification.body != null
+  //             ? Text(receivedNotification.body)
+  //             : null,
+  //         actions: [
+  //           CupertinoDialogAction(
+  //             isDefaultAction: true,
+  //             child: Text('Ok'),
+  //             onPressed: () async {
+  //               print('Not implemented Yet');
+  //             },
+  //           )
+  //         ],
+  //       ),
+  //     );
+  //   });
+  // }
 
-  void _configureSelectNotificationSubject() {
-    selectNotificationSubject.stream.listen((String payload) async {
-      print('Payload: $payload');
-      Map<String, dynamic> message = json.decode(payload);
-      String title = message['data']['title'];
-      String body = message['data']['body'];
-      onPushNotificationClick(title, body);
-    });
-  }
+  // void _configureSelectNotificationSubject() {
+  //   selectNotificationSubject.stream.listen((String payload) async {
+  //     print('Payload: $payload');
+  //     Map<String, dynamic> message = json.decode(payload);
+  //     String title = message['data']['title'];
+  //     String body = message['data']['body'];
+  //     onPushNotificationClick(title, body);
+  //   });
+  // }
 
-  void setupPushNotifications() async {
-    if (Platform.isIOS) {
-      _firebaseMessaging.requestNotificationPermissions(
-          IosNotificationSettings(sound: true, badge: true, alert: true));
-      _firebaseMessaging.onIosSettingsRegistered
-          .listen((IosNotificationSettings settings) {
-        print('Settings registered: $settings');
-      });
-    }
+  // void setupPushNotifications() async {
+  //   if (Platform.isIOS) {
+  //     await _firebaseMessaging.requestPermission(
+  //         sound: true, badge: true, alert: true);
+  //     // TODO: complete implementation as of new docs
+  //   }
 
-    _firebaseMessaging.configure(
-      onMessage: (Map<String, dynamic> message) async {
-        print('on message $message');
+  //   _firebaseMessaging. configure(
+  //     onMessage: (Map<String, dynamic> message) async {
+  //       print('on message $message');
 
-        //TODO Configure notification channel
-        var androidPlatformChannelSpecifics = AndroidNotificationDetails(
-            'silent channel id',
-            'silent channel name',
-            'silent channel description',
-            playSound: false,
-            styleInformation: DefaultStyleInformation(true, true));
-        var iOSPlatformChannelSpecifics =
-            IOSNotificationDetails(presentSound: false);
-        var platformChannelSpecifics = NotificationDetails(
-            androidPlatformChannelSpecifics, iOSPlatformChannelSpecifics);
-        await flutterLocalNotificationsPlugin.show(
-            0,
-            message['notification']['title'],
-            message['notification']['body'],
-            platformChannelSpecifics,
-            payload: json.encode(message));
-      },
-      onResume: (Map<String, dynamic> message) async {
-        print('on resume $message');
-        String title = message['data']['title'];
-        String body = message['data']['body'];
-        onPushNotificationClick(title, body);
-      },
-      onLaunch: (Map<String, dynamic> message) async {
-        print('on launch $message');
-        String title = message['data']['title'];
-        String body = message['data']['body'];
-        onPushNotificationClick(title, body);
-      },
-    );
+  //       //TODO Configure notification channel
+  //       var androidPlatformChannelSpecifics = AndroidNotificationDetails(
+  //           'silent channel id',
+  //           'silent channel name',
+  //           'silent channel description',
+  //           playSound: false,
+  //           styleInformation: DefaultStyleInformation(true, true));
+  //       var iOSPlatformChannelSpecifics =
+  //           IOSNotificationDetails(presentSound: false);
+  //       var platformChannelSpecifics = NotificationDetails(
+  //           androidPlatformChannelSpecifics, iOSPlatformChannelSpecifics);
+  //       await flutterLocalNotificationsPlugin.show(
+  //           0,
+  //           message['notification']['title'],
+  //           message['notification']['body'],
+  //           platformChannelSpecifics,
+  //           payload: json.encode(message));
+  //     },
+  //     onResume: (Map<String, dynamic> message) async {
+  //       print('on resume $message');
+  //       String title = message['data']['title'];
+  //       String body = message['data']['body'];
+  //       onPushNotificationClick(title, body);
+  //     },
+  //     onLaunch: (Map<String, dynamic> message) async {
+  //       print('on launch $message');
+  //       String title = message['data']['title'];
+  //       String body = message['data']['body'];
+  //       onPushNotificationClick(title, body);
+  //     },
+  //   );
 
-    await _firebaseMessaging.subscribeToTopic('announcements');
-  }
+  //   await _firebaseMessaging.subscribeToTopic('announcements');
+  // }
 
-  void onPushNotificationClick(String title, String body) async {
-    await showDialog(
-        context: context,
-        builder: (BuildContext context) {
-          return NotificationOnClickDialog(title: title, body: body,);
-        });
-  }
+  // void onPushNotificationClick(String title, String body) async {
+  //   await showDialog(
+  //       context: context,
+  //       builder: (BuildContext context) {
+  //         return NotificationOnClickDialog(
+  //           title: title,
+  //           body: body,
+  //         );
+  //       });
+  // }
 
   /// =================================================
   ///                SHOW TIMER BANNER
@@ -162,7 +161,7 @@ class DashboardState extends State<Dashboard> {
         child: TimerText(),
       ),
       actions: [
-        FlatButton(
+        TextButton(
           onPressed: () {
             setState(() {
               _displayTimerBanner = false;
@@ -199,14 +198,16 @@ class DashboardState extends State<Dashboard> {
         }
       }).then((networkSlacks) {
         if (networkSlacks != null) {
-          streamCtrl.sink.add(networkSlacks);
-          persistSlackAnnouncements(networkSlacks);
+          var announcements = networkSlacks as List<Announcement>;
+          streamCtrl.sink.add(announcements);
+          persistSlackAnnouncements(announcements);
           cacheTTL = DateTime.now().add(Duration(minutes: 5));
           streamCtrl.close();
         }
       });
     } catch (e) {
-      print('***********************\nSlack data stream ctrl error: ' + e);
+      print('***********************\nSlack data stream ctrl error: ' +
+          e.toString());
     }
     return streamCtrl.stream;
   }
@@ -237,7 +238,7 @@ class DashboardState extends State<Dashboard> {
               );
             default:
               print('ERROR-->DASHBOARD: ${snapshot.hasError}');
-              var resources = snapshot.data;
+              var resources = snapshot.data!;
               return Padding(
                 padding: EdgeInsets.symmetric(
                   horizontal: 10.0,
@@ -304,7 +305,7 @@ class DashboardState extends State<Dashboard> {
 /// ===============================================
 
 class TimerText extends StatefulWidget {
-  const TimerText({Key key}) : super(key: key);
+  const TimerText({Key? key}) : super(key: key);
 
   @override
   _TimerTextState createState() => _TimerTextState();
@@ -312,7 +313,7 @@ class TimerText extends StatefulWidget {
 
 class _TimerTextState extends State<TimerText> {
   DateTime _dateTime = DateTime.now();
-  Timer _timer;
+  Timer? _timer;
 
   @override
   void initState() {
