@@ -6,9 +6,9 @@ const Duration _kExpand = Duration(milliseconds: 200);
 
 class CustomExpansionTile extends StatefulWidget {
   const CustomExpansionTile({
-    Key key,
+    Key? key,
     this.leading,
-    @required this.title,
+    this.title,
     this.backgroundColor,
     this.onExpansionChanged,
     this.children = const <Widget>[],
@@ -17,13 +17,13 @@ class CustomExpansionTile extends StatefulWidget {
   })  : assert(initiallyExpanded != null),
         super(key: key);
 
-  final Widget leading;
-  final Widget title;
-  final ValueChanged<bool> onExpansionChanged;
-  final List<Widget> children;
-  final Color backgroundColor;
-  final Widget trailing;
-  final bool initiallyExpanded;
+  final Widget? leading;
+  final Widget? title;
+  final ValueChanged<bool>? onExpansionChanged;
+  final List<Widget>? children;
+  final Color? backgroundColor;
+  final Widget? trailing;
+  final bool? initiallyExpanded;
 
   @override
   CustomExpansionTileState createState() => CustomExpansionTileState();
@@ -31,14 +31,14 @@ class CustomExpansionTile extends StatefulWidget {
 
 class CustomExpansionTileState extends State<CustomExpansionTile>
     with SingleTickerProviderStateMixin {
-  AnimationController _controller;
-  CurvedAnimation _easeOutAnimation;
-  CurvedAnimation _easeInAnimation;
-  ColorTween _borderColor;
-  ColorTween _headerColor;
-  ColorTween _iconColor;
-  ColorTween _backgroundColor;
-  Animation<double> _iconTurns;
+  AnimationController? _controller;
+  CurvedAnimation? _easeOutAnimation;
+  CurvedAnimation? _easeInAnimation;
+  ColorTween? _borderColor;
+  ColorTween? _headerColor;
+  ColorTween? _iconColor;
+  ColorTween? _backgroundColor;
+  Animation<double>? _iconTurns;
 
   bool _isExpanded = false;
 
@@ -47,23 +47,23 @@ class CustomExpansionTileState extends State<CustomExpansionTile>
     super.initState();
     _controller = AnimationController(duration: _kExpand, vsync: this);
     _easeOutAnimation =
-        CurvedAnimation(parent: _controller, curve: Curves.easeOut);
+        CurvedAnimation(parent: _controller!, curve: Curves.easeOut);
     _easeInAnimation =
-        CurvedAnimation(parent: _controller, curve: Curves.easeIn);
+        CurvedAnimation(parent: _controller!, curve: Curves.easeIn);
     _borderColor = ColorTween();
     _headerColor = ColorTween();
     _iconColor = ColorTween();
-    _iconTurns = Tween<double>(begin: 0.0, end: 0.5).animate(_easeInAnimation);
+    _iconTurns = Tween<double>(begin: 0.0, end: 0.5).animate(_easeInAnimation!);
     _backgroundColor = ColorTween();
 
     _isExpanded =
         PageStorage.of(context)?.readState(context) ?? widget.initiallyExpanded;
-    if (_isExpanded) _controller.value = 1.0;
+    if (_isExpanded) _controller?.value = 1.0;
   }
 
   @override
   void dispose() {
-    _controller.dispose();
+    _controller?.dispose();
     super.dispose();
   }
 
@@ -84,9 +84,9 @@ class CustomExpansionTileState extends State<CustomExpansionTile>
       setState(() {
         _isExpanded = isExpanded;
         if (_isExpanded) {
-          _controller.forward();
+          _controller?.forward();
         } else {
-          _controller.reverse().then<void>((void value) {
+          _controller?.reverse().then<void>((void value) {
             if (!mounted) return;
             setState(() {
               // Rebuild without widget.children.
@@ -96,19 +96,19 @@ class CustomExpansionTileState extends State<CustomExpansionTile>
         PageStorage.of(context)?.writeState(context, _isExpanded);
       });
       if (widget.onExpansionChanged != null) {
-        widget.onExpansionChanged(_isExpanded);
+        widget.onExpansionChanged!(_isExpanded);
       }
     }
   }
 
-  Widget _buildChildren(BuildContext context, Widget child) {
+  Widget _buildChildren(BuildContext context, Widget? child) {
     final borderSideColor =
-        _borderColor.evaluate(_easeOutAnimation) ?? Colors.transparent;
-    final titleColor = _headerColor.evaluate(_easeInAnimation);
+        _borderColor?.evaluate(_easeOutAnimation!) ?? Colors.transparent;
+    final titleColor = _headerColor?.evaluate(_easeInAnimation!);
 
     return Container(
       decoration: BoxDecoration(
-          color: _backgroundColor.evaluate(_easeOutAnimation) ??
+          color: _backgroundColor?.evaluate(_easeOutAnimation!) ??
               Colors.transparent,
           border: Border(
             top: BorderSide(color: borderSideColor),
@@ -118,27 +118,23 @@ class CustomExpansionTileState extends State<CustomExpansionTile>
         mainAxisSize: MainAxisSize.min,
         children: <Widget>[
           IconTheme.merge(
-            data: IconThemeData(color: _iconColor.evaluate(_easeInAnimation)),
+            data: IconThemeData(color: _iconColor?.evaluate(_easeInAnimation!)),
             child: ListTile(
               onTap: toggle,
               leading: widget.leading,
               title: DefaultTextStyle(
                 style: Theme.of(context)
                     .textTheme
-                    .subtitle1
+                    .subtitle1!
                     .copyWith(color: titleColor),
-                child: widget.title,
+                child: widget.title!,
               ),
-              trailing: widget.trailing ??
-                  RotationTransition(
-                    turns: _iconTurns,
-                    child: const Icon(Icons.expand_more),
-                  ),
+              trailing: widget.trailing,
             ),
           ),
           ClipRect(
             child: Align(
-              heightFactor: _easeInAnimation.value,
+              heightFactor: _easeInAnimation?.value,
               child: child,
             ),
           ),
@@ -150,20 +146,18 @@ class CustomExpansionTileState extends State<CustomExpansionTile>
   @override
   Widget build(BuildContext context) {
     final theme = Theme.of(context);
-    _borderColor.end = theme.dividerColor;
-    _headerColor
-      ..begin = theme.textTheme.subtitle1.color
-      ..end = theme.accentColor;
-    _iconColor
-      ..begin = theme.unselectedWidgetColor
-      ..end = theme.accentColor;
-    _backgroundColor.end = widget.backgroundColor;
+    _borderColor?.end = theme.dividerColor;
+    _headerColor?.begin = theme.textTheme.subtitle1?.color;
+    // ?.end = theme.accentColor;
+    _iconColor?.begin = theme.unselectedWidgetColor;
+    // ..end = theme.accentColor;
+    _backgroundColor?.end = widget.backgroundColor;
 
-    final closed = !_isExpanded && _controller.isDismissed;
+    final closed = !_isExpanded && _controller!.isDismissed;
     return AnimatedBuilder(
-      animation: _controller.view,
+      animation: _controller!.view,
       builder: _buildChildren,
-      child: closed ? null : Column(children: widget.children),
+      child: closed ? null : Column(children: widget.children!),
     );
   }
 }
