@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import 'package:hidden_drawer_menu/hidden_drawer_menu.dart';
 import 'package:hidden_drawer_menu/simple_hidden_drawer/animated_drawer_content.dart';
 import 'package:hidden_drawer_menu/simple_hidden_drawer/simple_hidden_drawer.dart';
@@ -7,71 +8,71 @@ import 'custom_hidden_menu.dart';
 
 class CustomHiddenDrawerMenu extends StatelessWidget {
   /// List item menu and respective screens
-  final List<ScreenHiddenDrawer> screens;
+  final List<ScreenHiddenDrawer>? screens;
 
   /// position initial item selected in menu( sart in 0)
-  final int initPositionSelected;
+  final int? initPositionSelected;
 
   /// Decocator that allows us to add backgroud in the content(color)
-  final Color backgroundColorContent;
+  final Color? backgroundColorContent;
 
   //AppBar
   /// enable auto title in appbar with menu item name
-  final bool whithAutoTittleName;
+  final bool? whithAutoTittleName;
 
   /// Style of the title in appbar
-  final TextStyle styleAutoTittleName;
+  final TextStyle? styleAutoTittleName;
 
   /// change backgroundColor of the AppBar
-  final Color backgroundColorAppBar;
+  final Color? backgroundColorAppBar;
 
   ///Change elevation of the AppBar
-  final double elevationAppBar;
+  final double? elevationAppBar;
 
   ///Change leading of the AppBar
-  final Widget leadingAppBar;
+  final Widget? leadingAppBar;
 
   /// Add actions in the AppBar
-  final List<Widget> actionsAppBar;
+  final List<Widget>? actionsAppBar;
 
   /// Set custom widget in tittleAppBar
-  final Widget tittleAppBar;
+  final Widget? tittleAppBar;
 
   /// Decide whether title is centered or not
-  final bool isTitleCentered;
+  final bool? isTitleCentered;
 
   //Menu
   ///==== Custom Widget Allowed
-  final Widget backgroundMenu;
+  final Widget? backgroundMenu;
 
   /// that allows us to add backgroud in the menu(color)
-  final Color backgroundColorMenu;
+  final Color? backgroundColorMenu;
 
   /// that allows us to add shadow above menu items
-  final bool enableShadowItensMenu;
+  final bool? enableShadowItensMenu;
 
   /// enable and disable open and close with gesture
-  final bool isDraggable;
+  final bool? isDraggable;
 
-  final Curve curveAnimation;
+  final Curve? curveAnimation;
 
-  final double slidePercent;
+  final double? slidePercent;
 
   /// percent the content should scale vertically
-  final double verticalScalePercent;
+  final double? verticalScalePercent;
 
   /// radius applied to the content when active
-  final double contentCornerRadius;
+  final double? contentCornerRadius;
 
   /// enable animation Scale
-  final bool enableScaleAnimation;
+  final bool? enableScaleAnimation;
 
   /// enable animation borderRadius
-  final bool enableCornerAnimation;
+  final bool? enableCornerAnimation;
 
-  final bool disableAppBarDefault;
+  final bool? disableAppBarDefault;
 
-  final TypeOpen typeOpen;
+  final TypeOpen? typeOpen;
 
   CustomHiddenDrawerMenu({
     this.screens,
@@ -102,16 +103,16 @@ class CustomHiddenDrawerMenu extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return SimpleHiddenDrawer(
-      isDraggable: isDraggable,
-      curveAnimation: curveAnimation,
-      slidePercent: slidePercent,
-      verticalScalePercent: verticalScalePercent,
-      contentCornerRadius: contentCornerRadius,
-      enableCornerAnimation: enableCornerAnimation,
-      enableScaleAnimation: enableScaleAnimation,
+      isDraggable: isDraggable!,
+      curveAnimation: curveAnimation!,
+      slidePercent: slidePercent!,
+      verticalScalePercent: verticalScalePercent!,
+      contentCornerRadius: contentCornerRadius!,
+      enableCornerAnimation: enableCornerAnimation!,
+      enableScaleAnimation: enableScaleAnimation!,
       menu: buildMenu(),
-      typeOpen: typeOpen,
-      initPositionSelected: initPositionSelected,
+      typeOpen: typeOpen!,
+      initPositionSelected: initPositionSelected!,
       screenSelectedBuilder: (position, bloc) {
         // var actions = <Widget>[];
         // if (typeOpen == TypeOpen.FROM_RIGHT) {
@@ -128,7 +129,7 @@ class CustomHiddenDrawerMenu extends StatelessWidget {
         return Scaffold(
           backgroundColor: backgroundColorContent,
           appBar: AppBar(
-            brightness: Brightness.light,
+            systemOverlayStyle: SystemUiOverlayStyle.light,
             backgroundColor: backgroundColorAppBar,
             elevation: elevationAppBar,
             title: getTittleAppBar(position),
@@ -136,7 +137,7 @@ class CustomHiddenDrawerMenu extends StatelessWidget {
             leading: _buildLeading(bloc),
             actions: actionsAppBar,
           ),
-          body: screens[position].screen,
+          body: screens![position].screen,
         );
       },
     );
@@ -144,37 +145,37 @@ class CustomHiddenDrawerMenu extends StatelessWidget {
 
   Widget getTittleAppBar(int position) {
     if (tittleAppBar == null) {
-      return whithAutoTittleName
+      return whithAutoTittleName!
           ? Text(
-              screens[position].itemMenu.name,
+              screens![position].itemMenu.name,
               style: styleAutoTittleName,
             )
           : Container();
     } else {
-      return tittleAppBar;
+      return tittleAppBar!;
     }
   }
 
   Widget buildMenu() {
     var _itensMenu = <ItemHiddenMenu>[];
 
-    screens.forEach((item) {
+    screens?.forEach((item) {
       _itensMenu.add(item.itemMenu);
     });
 
     return CustomHiddenMenu(
       itens: _itensMenu,
-      background: backgroundMenu,
-      backgroundColorMenu: backgroundColorMenu,
-      initPositionSelected: initPositionSelected,
-      enableShadowItensMenu: enableShadowItensMenu,
-      typeOpen: typeOpen,
+      background: backgroundMenu!,
+      backgroundColorMenu: backgroundColorMenu!,
+      initPositionSelected: initPositionSelected!,
+      enableShadowItensMenu: enableShadowItensMenu!,
+      typeOpen: typeOpen!,
     );
   }
 
   Widget _buildLeading(SimpleHiddenDrawerController bloc) {
     if (typeOpen == TypeOpen.FROM_LEFT) {
-      return IconButton(icon: leadingAppBar, onPressed: () => bloc.toggle());
+      return IconButton(icon: leadingAppBar!, onPressed: () => bloc.toggle());
     } else {
       return SizedBox.shrink();
     }
