@@ -22,31 +22,31 @@ typedef Widget DigitBuilder(BuildContext, int);
 ///
 @immutable
 class FlipClock extends StatelessWidget {
-  DigitBuilder _digitBuilder;
-  Widget _separator;
-  final DateTime startTime;
-  final EdgeInsets spacing;
-  final FlipDirection flipDirection;
+  DigitBuilder? _digitBuilder;
+  Widget? _separator;
+  final DateTime? startTime;
+  final EdgeInsets? spacing;
+  final FlipDirection? flipDirection;
 
   /// Set countdown to true to have a countdown timer.
-  final bool countdownMode;
+  final bool? countdownMode;
 
-  final bool _showHours;
-  final bool _showDays;
+  final bool? _showHours;
+  final bool? _showDays;
 
-  Duration timeLeft;
+  Duration? timeLeft;
 
   /// Called when the countdown clock hits zero.
-  final VoidCallback onDone;
+  final VoidCallback? onDone;
 
-  final double height;
-  final double width;
+  final double? height;
+  final double? width;
 
   FlipClock({
-    Key key,
-    @required DigitBuilder digitBuilder,
-    @required Widget separator,
-    @required this.startTime,
+    Key? key,
+    required DigitBuilder digitBuilder,
+    required Widget separator,
+    required this.startTime,
     this.countdownMode = false,
     this.spacing = const EdgeInsets.symmetric(horizontal: 2.0),
     this.flipDirection = FlipDirection.up,
@@ -60,11 +60,11 @@ class FlipClock extends StatelessWidget {
         onDone = null;
 
   FlipClock.simple({
-    Key key,
-    @required this.startTime,
-    @required Color digitColor,
-    @required Color backgroundColor,
-    @required double digitSize,
+    Key? key,
+    required this.startTime,
+    required Color digitColor,
+    required Color backgroundColor,
+    required double digitSize,
     BorderRadius borderRadius = const BorderRadius.all(Radius.circular(0.0)),
     this.spacing = const EdgeInsets.symmetric(horizontal: 2.0),
     this.flipDirection = FlipDirection.up,
@@ -96,7 +96,7 @@ class FlipClock extends StatelessWidget {
         color: backgroundColor,
         borderRadius: borderRadius,
       ),
-      width: width / 2,
+      width: width! / 2,
       height: height,
       alignment: Alignment.center,
       child: Text(
@@ -110,11 +110,11 @@ class FlipClock extends StatelessWidget {
   }
 
   FlipClock.countdown({
-    Key key,
-    @required Duration duration,
-    @required Color digitColor,
-    @required Color backgroundColor,
-    @required double digitSize,
+    Key? key,
+    required Duration duration,
+    required Color digitColor,
+    required Color backgroundColor,
+    required double digitSize,
     BorderRadius borderRadius = const BorderRadius.all(Radius.circular(0.0)),
     this.spacing = const EdgeInsets.symmetric(horizontal: 2.0),
     this.onDone,
@@ -147,7 +147,7 @@ class FlipClock extends StatelessWidget {
         color: backgroundColor,
         borderRadius: borderRadius,
       ),
-      width: width / 2,
+      width: width! / 2,
       height: height,
       alignment: Alignment.center,
       child: Text(
@@ -161,13 +161,13 @@ class FlipClock extends StatelessWidget {
   }
 
   FlipClock.reverseCountdown({
-    Key key,
+    Key? key,
     // @required DateTime now,
     // @required DateTime dDay,
-    @required Duration duration,
-    @required Color digitColor,
-    @required Color backgroundColor,
-    @required double digitSize,
+    required Duration duration,
+    required Color digitColor,
+    required Color backgroundColor,
+    required double digitSize,
     BorderRadius borderRadius = const BorderRadius.all(Radius.circular(0.0)),
     this.spacing = const EdgeInsets.symmetric(horizontal: 2.0),
     this.onDone,
@@ -200,7 +200,7 @@ class FlipClock extends StatelessWidget {
         color: backgroundColor,
         borderRadius: borderRadius,
       ),
-      width: width / 2,
+      width: width! / 2,
       height: height,
       alignment: Alignment.center,
       child: Text(
@@ -219,42 +219,42 @@ class FlipClock extends StatelessWidget {
     final initStream =
         Stream<DateTime>.periodic(Duration(milliseconds: 1000), (_) {
       var oldTime = time;
-      (countdownMode)
-          ? timeLeft = timeLeft - Duration(seconds: 1)
-          : time = time.add(Duration(seconds: 1));
+      (countdownMode!)
+          ? timeLeft = timeLeft! - Duration(seconds: 1)
+          : time = time?.add(Duration(seconds: 1));
 
-      if (!countdownMode && oldTime.day != time.day) {
+      if (!countdownMode! && oldTime?.day != time?.day) {
         time = oldTime;
-        if (onDone != null) onDone();
+        if (onDone != null) onDone!();
       }
 
-      return time;
+      return time!;
     });
     final timeStream =
-        (countdownMode ? initStream.take(timeLeft.inSeconds) : initStream)
+        (countdownMode! ? initStream.take(timeLeft!.inSeconds) : initStream)
             .asBroadcastStream();
 
     var digitList = <Widget>[];
     // TODO(efortuna): Instead, allow the user to specify the format of time instead.
     // Add hours if appropriate.
 
-    if (_showDays) {
+    if (_showDays!) {
       digitList.addAll([
         _buildSegment(
             timeStream,
             (DateTime time) =>
-                (timeLeft.inDays > 99) ? 9 : (timeLeft.inDays ~/ 10),
+                (timeLeft!.inDays > 99) ? 9 : (timeLeft!.inDays ~/ 10),
             (DateTime time) =>
-                (timeLeft.inDays > 99) ? 9 : (timeLeft.inDays % 10),
-            startTime,
+                (timeLeft!.inDays > 99) ? 9 : (timeLeft!.inDays % 10),
+            startTime!,
             "days"),
         Column(
           children: <Widget>[
             Padding(
-              padding: spacing,
+              padding: spacing!,
               child: _separator,
             ),
-            (_showDays)
+            (_showDays!)
                 ? Container(color: Colors.black)
                 : Container(
                     color: Colors.transparent,
@@ -264,25 +264,25 @@ class FlipClock extends StatelessWidget {
       ]);
     }
 
-    if (_showHours) {
+    if (_showHours!) {
       digitList.addAll([
         _buildSegment(
             timeStream,
-            (DateTime time) => (countdownMode)
-                ? (timeLeft.inHours % 24) ~/ 10
+            (DateTime time) => (countdownMode!)
+                ? (timeLeft!.inHours % 24) ~/ 10
                 : (time.hour) ~/ 10,
-            (DateTime time) => (countdownMode)
-                ? (timeLeft.inHours % 24) % 10
+            (DateTime time) => (countdownMode!)
+                ? (timeLeft!.inHours % 24) % 10
                 : (time.hour) % 10,
-            startTime,
+            startTime!,
             "Hours"),
         Column(
           children: <Widget>[
             Padding(
-              padding: spacing,
+              padding: spacing!,
               child: _separator,
             ),
-            (_showDays)
+            (_showDays!)
                 ? Container(color: Colors.black)
                 : Container(
                     color: Colors.transparent,
@@ -300,22 +300,22 @@ class FlipClock extends StatelessWidget {
             // Minutes
             _buildSegment(
                 timeStream,
-                (DateTime time) => (countdownMode)
-                    ? (timeLeft.inMinutes % 60) ~/ 10
+                (DateTime time) => (countdownMode!)
+                    ? (timeLeft!.inMinutes % 60) ~/ 10
                     : (time.minute) ~/ 10,
-                (DateTime time) => (countdownMode)
-                    ? (timeLeft.inMinutes % 60) % 10
+                (DateTime time) => (countdownMode!)
+                    ? (timeLeft!.inMinutes % 60) % 10
                     : (time.minute) % 10,
-                startTime,
+                startTime!,
                 "minutes"),
 
             Column(
               children: <Widget>[
                 Padding(
-                  padding: spacing,
+                  padding: spacing!,
                   child: _separator,
                 ),
-                (_showDays)
+                (_showDays!)
                     ? Container(color: Colors.black)
                     : Container(
                         color: Colors.transparent,
@@ -326,46 +326,46 @@ class FlipClock extends StatelessWidget {
             // Seconds
             _buildSegment(
                 timeStream,
-                (DateTime time) => (countdownMode)
-                    ? (timeLeft.inSeconds % 60) ~/ 10
+                (DateTime time) => (countdownMode!)
+                    ? (timeLeft!.inSeconds % 60) ~/ 10
                     : (time.second) ~/ 10,
-                (DateTime time) => (countdownMode)
-                    ? (timeLeft.inSeconds % 60) % 10
+                (DateTime time) => (countdownMode!)
+                    ? (timeLeft!.inSeconds % 60) % 10
                     : (time.second) % 10,
-                startTime,
+                startTime!,
                 "seconds")
           ]),
       ),
     );
   }
 
-  _buildSegment(Stream<DateTime> timeStream, Function tensDigit,
-      Function onesDigit, DateTime startTime, String id) {
+  _buildSegment(Stream<DateTime> timeStream, int Function(DateTime) tensDigit,
+      int Function(DateTime) onesDigit, DateTime startTime, String id) {
     return Column(
       children: <Widget>[
         Row(children: [
           Padding(
-            padding: spacing,
+            padding: spacing!,
             child: FlipPanel<int>.stream(
               itemStream: timeStream.map<int>(tensDigit),
-              itemBuilder: _digitBuilder,
+              itemBuilder: _digitBuilder!,
               duration: const Duration(milliseconds: 450),
               initValue: tensDigit(startTime),
-              direction: flipDirection,
+              direction: flipDirection!,
             ),
           ),
           Padding(
-            padding: spacing,
+            padding: spacing!,
             child: FlipPanel<int>.stream(
               itemStream: timeStream.map<int>(onesDigit),
-              itemBuilder: _digitBuilder,
+              itemBuilder: _digitBuilder!,
               duration: const Duration(milliseconds: 450),
               initValue: onesDigit(startTime),
-              direction: flipDirection,
+              direction: flipDirection!,
             ),
           ),
         ]),
-        (_showDays)
+        (_showDays!)
             ? Row(
                 children: <Widget>[
                   Padding(
@@ -422,20 +422,20 @@ enum FlipDirection { up, down }
 /// a static [Widget] (which is an instance of [StatelessWidget])
 ///
 class FlipPanel<T> extends StatefulWidget {
-  final IndexedItemBuilder indexedItemBuilder;
-  final StreamItemBuilder<T> streamItemBuilder;
-  final Stream<T> itemStream;
-  final int itemsCount;
-  final Duration period;
-  final Duration duration;
-  final int loop;
-  final int startIndex;
-  final T initValue;
-  final double spacing;
-  final FlipDirection direction;
+  final IndexedItemBuilder? indexedItemBuilder;
+  final StreamItemBuilder<T>? streamItemBuilder;
+  final Stream<T>? itemStream;
+  final int? itemsCount;
+  final Duration? period;
+  final Duration? duration;
+  final int? loop;
+  final int? startIndex;
+  final T? initValue;
+  final double? spacing;
+  final FlipDirection? direction;
 
   FlipPanel({
-    Key key,
+    Key? key,
     this.indexedItemBuilder,
     this.streamItemBuilder,
     this.itemStream,
@@ -459,8 +459,8 @@ class FlipPanel<T> extends StatefulWidget {
   /// otherwise the animation becomes jerky/stuttery.
   ///
   FlipPanel.builder({
-    Key key,
-    @required IndexedItemBuilder itemBuilder,
+    Key? key,
+    @required IndexedItemBuilder? itemBuilder,
     @required this.itemsCount,
     @required this.period,
     this.duration = const Duration(milliseconds: 500),
@@ -470,9 +470,9 @@ class FlipPanel<T> extends StatefulWidget {
     this.direction = FlipDirection.up,
   })  : assert(itemBuilder != null),
         assert(itemsCount != null),
-        assert(startIndex < itemsCount),
+        assert(startIndex! < itemsCount!),
         assert(period == null ||
-            period.inMilliseconds >= 2 * duration.inMilliseconds),
+            period.inMilliseconds >= 2 * duration!.inMilliseconds),
         indexedItemBuilder = itemBuilder,
         streamItemBuilder = null,
         itemStream = null,
@@ -485,9 +485,9 @@ class FlipPanel<T> extends StatefulWidget {
   /// * [itemBuilder] is called whenever a new value is emitted from [itemStream]
   ///
   FlipPanel.stream({
-    Key key,
+    Key? key,
     @required this.itemStream,
-    @required StreamItemBuilder<T> itemBuilder,
+    @required StreamItemBuilder<T>? itemBuilder,
     this.initValue,
     this.duration = const Duration(milliseconds: 500),
     this.spacing = 0.5,
@@ -507,23 +507,23 @@ class FlipPanel<T> extends StatefulWidget {
 
 class _FlipPanelState<T> extends State<FlipPanel>
     with SingleTickerProviderStateMixin {
-  AnimationController _controller;
-  Animation _animation;
-  int _currentIndex;
-  bool _isReversePhase;
-  bool _isStreamMode;
-  bool _running;
+  AnimationController? _controller;
+  Animation? _animation;
+  int? _currentIndex;
+  bool? _isReversePhase;
+  bool? _isStreamMode;
+  bool? _running;
   final _perspective = 0.003;
   final _zeroAngle =
       0.0001; // There's something wrong in the perspective transform, I use a very small value instead of zero to temporarily get it around.
-  int _loop;
-  T _currentValue, _nextValue;
-  Timer _timer;
-  StreamSubscription<T> _subscription;
+  int? _loop;
+  T? _currentValue, _nextValue;
+  Timer? _timer;
+  StreamSubscription<T>? _subscription;
 
-  Widget _child1, _child2;
-  Widget _upperChild1, _upperChild2;
-  Widget _lowerChild1, _lowerChild2;
+  Widget? _child1, _child2;
+  Widget? _upperChild1, _upperChild2;
+  Widget? _lowerChild1, _lowerChild2;
 
   @override
   void initState() {
@@ -539,7 +539,7 @@ class _FlipPanelState<T> extends State<FlipPanel>
           ..addStatusListener((status) {
             if (status == AnimationStatus.completed) {
               _isReversePhase = true;
-              _controller.reverse();
+              _controller?.reverse();
             }
             if (status == AnimationStatus.dismissed) {
               _currentValue = _nextValue;
@@ -552,21 +552,21 @@ class _FlipPanelState<T> extends State<FlipPanel>
             });
           });
     _animation =
-        Tween(begin: _zeroAngle, end: math.pi / 2).animate(_controller);
+        Tween(begin: _zeroAngle, end: math.pi / 2).animate(_controller!);
 
     if (widget.period != null) {
-      _timer = Timer.periodic(widget.period, (_) {
-        if (widget.loop < 0 || _loop < widget.loop) {
-          if (_currentIndex + 1 == widget.itemsCount - 2) {
-            _loop++;
+      _timer = Timer.periodic(widget.period!, (_) {
+        if (widget.loop! < 0 || _loop! < widget.loop!) {
+          if (_currentIndex! + 1 == widget.itemsCount! - 2) {
+            _loop = _loop! + 1;
           }
-          _currentIndex = (_currentIndex + 1) % widget.itemsCount;
+          _currentIndex = (_currentIndex! + 1) % widget.itemsCount!;
           _child1 = null;
           _isReversePhase = false;
-          _controller.forward();
+          _controller?.forward();
         } else {
-          _timer.cancel();
-          _currentIndex = (_currentIndex + 1) % widget.itemsCount;
+          _timer?.cancel();
+          _currentIndex = (_currentIndex! + 1) % widget.itemsCount!;
           setState(() {
             _running = false;
           });
@@ -574,28 +574,28 @@ class _FlipPanelState<T> extends State<FlipPanel>
       });
     }
 
-    if (_isStreamMode) {
+    if (_isStreamMode!) {
       _currentValue = widget.initValue;
-      _subscription = widget.itemStream.distinct().listen((value) {
+      _subscription = widget.itemStream?.distinct().listen((value) {
         if (_currentValue == null) {
           _currentValue = value;
         } else if (value != _currentValue) {
           _nextValue = value;
           _child1 = null;
           _isReversePhase = false;
-          _controller.forward();
+          _controller?.forward();
         }
-      });
-    } else if (widget.loop < 0 || _loop < widget.loop) {
-      _controller.forward();
+      }) as StreamSubscription<T>?;
+    } else if (widget.loop! < 0 || _loop! < widget.loop!) {
+      _controller?.forward();
     }
   }
 
   @override
   void dispose() {
-    _controller.dispose();
-    if (_subscription != null) _subscription.cancel();
-    if (_timer != null) _timer.cancel();
+    _controller?.dispose();
+    if (_subscription != null) _subscription?.cancel();
+    if (_timer != null) _timer?.cancel();
     super.dispose();
   }
 
@@ -627,39 +627,49 @@ class _FlipPanelState<T> extends State<FlipPanel>
       );
     }
 
-    if (_running) {
+    if (_running!) {
       if (_child1 == null) {
-        _child1 = _child2 != null
-            ? _child2
-            : _isStreamMode
-                ? widget.streamItemBuilder(context, _currentValue)
-                : widget.indexedItemBuilder(
-                    context, _currentIndex % widget.itemsCount);
+        if (_child2 != null) {
+          _child1 = _child2;
+        } else {
+          _child1 = _isStreamMode!
+              ? widget.streamItemBuilder!(context, _currentValue)
+              : widget.indexedItemBuilder!(
+                  context, _currentIndex! % widget.itemsCount!);
+        }
         _child2 = null;
-        _upperChild1 =
-            _upperChild2 != null ? _upperChild2 : makeUpperClip(_child1);
-        _lowerChild1 =
-            _lowerChild2 != null ? _lowerChild2 : makeLowerClip(_child1);
+        if (_upperChild2 != null) {
+          _upperChild1 = _upperChild2;
+        } else {
+          _upperChild1 = makeUpperClip(_child1!);
+        }
+        if (_lowerChild2 != null) {
+          _lowerChild1 = _lowerChild2;
+        } else {
+          _lowerChild1 = makeLowerClip(_child1!);
+        }
       }
       if (_child2 == null) {
-        _child2 = _isStreamMode
-            ? widget.streamItemBuilder(context, _nextValue)
-            : widget.indexedItemBuilder(
-                context, (_currentIndex + 1) % widget.itemsCount);
-        _upperChild2 = makeUpperClip(_child2);
-        _lowerChild2 = makeLowerClip(_child2);
+        _child2 = _isStreamMode!
+            ? widget.streamItemBuilder!(context, _nextValue)
+            : widget.indexedItemBuilder!(
+                context, (_currentIndex! + 1) % widget.itemsCount!);
+        _upperChild2 = makeUpperClip(_child2!);
+        _lowerChild2 = makeLowerClip(_child2!);
       }
     } else {
-      _child1 = _child2 != null
-          ? _child2
-          : _isStreamMode
-              ? widget.streamItemBuilder(context, _currentValue)
-              : widget.indexedItemBuilder(
-                  context, _currentIndex % widget.itemsCount);
+      if (_child2 != null) {
+        _child1 = _child2;
+      } else {
+        _child1 = _isStreamMode!
+            ? widget.streamItemBuilder!(context, _currentValue)
+            : widget.indexedItemBuilder!(
+                context, _currentIndex! % widget.itemsCount!);
+      }
       _upperChild1 =
-          _upperChild2 != null ? _upperChild2 : makeUpperClip(_child1);
+          _upperChild2 != null ? _upperChild2 : makeUpperClip(_child1!);
       _lowerChild1 =
-          _lowerChild2 != null ? _lowerChild2 : makeLowerClip(_child1);
+          _lowerChild2 != null ? _lowerChild2 : makeLowerClip(_child1!);
     }
   }
 
@@ -676,7 +686,7 @@ class _FlipPanelState<T> extends State<FlipPanel>
               alignment: Alignment.bottomCenter,
               transform: Matrix4.identity()
                 ..setEntry(3, 2, _perspective)
-                ..rotateX(_isReversePhase ? _animation.value : math.pi / 2),
+                ..rotateX(_isReversePhase! ? _animation!.value : math.pi / 2),
               child: _upperChild2,
             ),
           ],
@@ -693,7 +703,7 @@ class _FlipPanelState<T> extends State<FlipPanel>
               alignment: Alignment.bottomCenter,
               transform: Matrix4.identity()
                 ..setEntry(3, 2, _perspective)
-                ..rotateX(_isReversePhase ? math.pi / 2 : _animation.value),
+                ..rotateX(_isReversePhase! ? math.pi / 2 : _animation!.value),
               child: _upperChild1,
             ),
           ],
@@ -712,7 +722,7 @@ class _FlipPanelState<T> extends State<FlipPanel>
               alignment: Alignment.topCenter,
               transform: Matrix4.identity()
                 ..setEntry(3, 2, _perspective)
-                ..rotateX(_isReversePhase ? math.pi / 2 : -_animation.value),
+                ..rotateX(_isReversePhase! ? math.pi / 2 : -_animation!.value),
               child: _lowerChild1,
             )
           ],
@@ -729,14 +739,14 @@ class _FlipPanelState<T> extends State<FlipPanel>
               alignment: Alignment.topCenter,
               transform: Matrix4.identity()
                 ..setEntry(3, 2, _perspective)
-                ..rotateX(_isReversePhase ? -_animation.value : math.pi / 2),
+                ..rotateX(_isReversePhase! ? -_animation!.value : math.pi / 2),
               child: _lowerChild2,
             )
           ],
         );
 
   Widget _buildPanel() {
-    return _running
+    return _running!
         ? Column(
             mainAxisSize: MainAxisSize.min,
             mainAxisAlignment: MainAxisAlignment.center,
@@ -744,12 +754,12 @@ class _FlipPanelState<T> extends State<FlipPanel>
             children: [
               _buildUpperFlipPanel(),
               Padding(
-                padding: EdgeInsets.only(top: widget.spacing),
+                padding: EdgeInsets.only(top: widget.spacing!),
               ),
               _buildLowerFlipPanel(),
             ],
           )
-        : _isStreamMode && _currentValue == null
+        : _isStreamMode! && _currentValue == null
             ? Container()
             : Column(
                 mainAxisSize: MainAxisSize.min,
@@ -763,7 +773,7 @@ class _FlipPanelState<T> extends State<FlipPanel>
                         ..rotateX(_zeroAngle),
                       child: _upperChild1),
                   Padding(
-                    padding: EdgeInsets.only(top: widget.spacing),
+                    padding: EdgeInsets.only(top: widget.spacing!),
                   ),
                   Transform(
                       alignment: Alignment.topCenter,
