@@ -3,7 +3,6 @@ import 'package:HackRU/ui/pages/about_app/about.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
-import 'package:groovin_material_icons/groovin_material_icons.dart';
 
 void main() {
   testWidgets('About App Test', (WidgetTester tester) async {
@@ -13,7 +12,7 @@ void main() {
 
     final findWebButton = find.widgetWithIcon(
       SocialMediaCard,
-      GroovinMaterialIcons.web,
+      FontAwesomeIcons.link,
       skipOffstage: false,
     );
     final findGithubButton = find.widgetWithIcon(
@@ -37,9 +36,5 @@ void main() {
     expect(findGithubButton, findsOneWidget);
     expect(findFacebookButton, findsOneWidget);
     expect(findInstagramButton, findsOneWidget);
-
-    // await tester
-    //     .tap(find.widgetWithIcon(SocialMediaCard, GroovinMaterialIcons.web));
-    // await tester.pump();
   });
 }
